\name{biweight}
\alias{biweight}
\title{
Bivariate Boxplot Robust 
}
\description{
In bvbox when method = "robust" this function is used.
}
\usage{
biweight(a,const1=9,const2=36,err=0.0001)
}
\arguments{
  \item{a}{a matrix}
  \item{const1}{common tuning constant}
  \item{const2}{bivariate tuning constant}
  \item{err}{convergence criterion}
}
\details{
See Chapter 2: Visualization.
}
\keyword{bvbox}
