\name{bvbox}
\alias{bvbox}
\title{
Bivariate Boxplot
}
\description{
Boxplots in two dimensions
}
\usage{
bvbox(a, d = 7, mtitle = "Bivariate Boxplot", method = "robust", 
      xlab = "X", ylab = "Y", add = FALSE, ...)
}
\arguments{
  \item{a}{a matrix}
  \item{d}{dimension}
  \item{mtitle}{title}
  \item{method}{character}
  \item{xlab}{x axis label}
  \item{ylab}{y axis label}
  \item{add}{add to existing plot?}
  \item{\dots}{additional plot arguments}
}
\details{
See Chapter 2: Visualization.
}
\keyword{plot}
