\name{chiplot}
\alias{chiplot}
\title{Chi-squared Plot}
\description{
  Chi-squared plot for independence
}
\usage{
chiplot(x, y, ...)
}
\arguments{
  \item{x}{a numeric vector}
  \item{y}{a numeric vector}
  \item{\dots}{additional plot arguments}
}
\details{
See Chapter 2: Visualization.
}
\keyword{plot}
